Switch to hand-rolled memcpy for building inline strings (on 64-bit)
Previously this function contained a call to libc's memcpy, which
essentially resulted in a non-inlineable function call overhead for
calling Yarn::inlined.

On architectures where RawYarn is a u128 wide, we use "the usual"
binary search trick to load 1 to 15 bytes into a "register" at the cost
of 2-3 loads and 1-3 branches. Elsewhere we fall back to libc's
memcpy().

Did I benchmark this? Of course not.