gilded: make Test::outputs() return handles for each output
In the previous design of output(), you could get into a situation where
the outputs for the test depended on which output path the test took,
meaning that if an error occured, it would generate an error golden, but
once the error was resolved, it would fail to delete the error golden.
This change requires that all outputs be specified up-front.