gilded: Ensure that changes in test inputs are picked up
Rust does not provide a way to specify that there are non-Rust files
relevant to the build that should cause the build to stale if they
change. We implement this ourselves through some gross mtime hacks.

The best we can do is inform the user that inputs have changed and
they must re-run cargo test to cause the tests to be rebuilt against
the new inputs.